[ParkingLot] Implement ParkingLot
Design a parking lot with multiple floors.
The parking lot has diffrent parking spaces for 2 wheelers,
4 wheelers, inactive and Electric.
Assign an empty parking spot to a vehicle when vehicle enters
and remove the vehicle when it is unparked.
Each vehicle will have a vehicleNumber. The system should
provide a spot that is nearest to entry point floor.
The parking lot should have multiple entry and exit points.
Customers can collect a parking ticket from the entry points
and can pay the
parking fee at the exit points on their way out.
System should display the fare when exit.
The system should support a per-hour parking fee model.
$4 for the first hour $3.5 for the second and third hours,
and $2.5 for all the remaining hours.
Get the history of a vehicle parked for last 3 months