[ElevatorManagementSystem] add elevator manager
Add threads to run elevatros in real time
Add and remove elevators
Track state of lifts i.e. floor lift is
at, move direction, number of passengers,
existing and future requests. assign
lift optimally to a user who is on a given
floor and wants to go to a destination floor.